Comunicate variables through template inheritance
In Tera, block redefinition is semantically equivalent to _replacing_
the parent's block contents by the redefined ones.

That means that if we reserve a block to manage variables that are later
used outside of said block, we can effectively change their values from
a child template to be used by the parent.

This is how I'm going to enable inclusion of only the used components
from Fomantic-UI in each page while also avoiding (repeated) reinclusion.